requirements.txt
altair                    
asgiref                   
attrs                     
blinker                   
cachetools                
certifi                   
charset-normalizer        
click                     
colorama                  
contourpy                 
cycler                    
Django                    
django-cors-headers       
djangorestframework      
fonttools                 
gitdb                     
GitPython                 
idna                    
importlib_resources       
Jinja2                    
joblib                    
jsonschema                
jsonschema-specifications 
kiwisolver                
markdown-it-py            
MarkupSafe                
matplotlib                
mdurl                     
numpy                    
packaging                
pandas                    
pillow                    
pip                      
protobuf                 
pyarrow                   
pydeck                    
Pygments                  
pyparsing                 
python-dateutil           
pytz                      
referencing               
requests                  
rich                      
rpds-py                   
scikit-learn              
scipy                     
seaborn                   
setuptools                
six                       
smmap                     
sqlparse                  
streamlit                 
tenacity                  
threadpoolctl             
toml                      
toolz                     
tornado                 
typing_extensions         
tzdata                    
urllib3                   
watchdog                  
zipp                                